WIP move saved searches to DB, start on CLI

+++ db/migrate/20260912002700_create_saved_searches.rb
@@ -2,8 +2,8 @@ class CreateSavedSearches < ActiveRecord::Migration[8.1]
   def change
     create_table :saved_searches do |t|
       t.string :query
-      t.double :lat
-      t.double :long
+      t.decimal :lat
+      t.decimal :long
       t.decimal :min_price
       t.decimal :max_price
       t.integer :condition
